% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{all_fit}
\alias{all_fit}
\title{Function to fit list of models.}
\usage{
all_fit(sf, fns)
}
\arguments{
\item{sf}{A simple features data frame.}

\item{fns}{A list of fitting functions.}
}
\value{
A list of fitted models.
}
\description{
Function to fit list of models.
}
