% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bym2.R
\name{bym2_inla}
\alias{bym2_inla}
\title{Fit BYM2 Small Area Estimation model using \code{R-INLA}.}
\usage{
bym2_inla(sf, verbose = FALSE)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{verbose}{Should \code{R-INLA} run in mode \code{verbose = TRUE}.}
}
\description{
Fit BYM2 Small Area Estimation model using \code{R-INLA}.
}
\examples{
bym2_inla(mw)
}
