% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bym2.R
\name{bym2_stan}
\alias{bym2_stan}
\title{Fit BYM2 Small Area Estimation model using \code{stan}.}
\usage{
bym2_stan(sf, nsim_warm = 100, nsim_iter = 1000, method = "default")
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}
}
\description{
Fit BYM2 Small Area Estimation model using \code{stan}.
}
\examples{
bym2_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
