% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_structures.R
\name{centroid_distance}
\alias{centroid_distance}
\title{Compute distances between area centroids.}
\usage{
centroid_distance(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\value{
A \code{nrow(sf)} by \code{nrow(sf)} matrix.
}
\description{
Wrapper function for \code{sf::centroid} and \code{sf::distance}.
}
\examples{
centroid_distance(mw)
}
