% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{cross_validate}
\alias{cross_validate}
\title{Cross-validation using \code{\link{lobo}} and \code{\link{eval_inla_model}}.}
\usage{
cross_validate(sf, fitting_fun, S = 5000)
}
\arguments{
\item{sf}{A simple features data frame.}

\item{S}{The number of Monte Carlo samples to draw from the \code{R-INLA}
approximate posterior distribution over the latent field.}

\item{fiting_fun}{An \code{R-INLA} model fitting function such as
\code{m1_inla}.}
}
\value{
A vector of log density scores (one entry for training data set
produced by \code{\link{lobo}}).
}
\description{
Cross-validation using \code{\link{lobo}} and \code{\link{eval_inla_model}}.
}
\examples{
cross_validate(mw, fitting_fun = m1_inla)
}
