% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{cv}
\alias{cv}
\title{Cross-validation using \code{\link{create_folds}}.}
\usage{
cv(sf, type = "LOO", fn, ..., S = 5000, return_fits = FALSE)
}
\arguments{
\item{sf}{A simple features data frame.}

\item{type}{One of \code{"LOO"} or \code{"SLOO"}.}

\item{fn}{A model fitting function.}

\item{...}{Additional arguments to \code{fn}.}

\item{S}{The number of Monte Carlo samples to draw from the \code{R-INLA}
approximate posterior distribution over the latent field.}

\item{return_fits}{Should \code{cv} return the training data and fitted models? Defaults to \code{FALSE} in case of low storage capacity.}
}
\value{
A vector of log density scores (one entry for training data set
produced by \code{\link{create_folds}}.
}
\description{
This function fits an model specified via \code{fn} upon each of
the training datasets generated by \code{create_folds} and then evaluates
their prediction according to either \code{eval_stan_model} or
\code{eval_inla_model}.
}
\examples{
cv(mw, type = "LOO", fn = m1_inla)
}
