% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{eval_inla_model}
\alias{eval_inla_model}
\title{Compute log density score from `R-INLA} model at a single held-out
point.}
\usage{
eval_inla_model(sf, fit, i, S = 5000)
}
\arguments{
\item{sf}{A simple features data frame.}

\item{fit}{A fitted \code{R-INLA} model.}

\item{i}{The index of the held-out data point to predict on. Should be an
integer in the range \code{1:nrow(sf)}.}

\item{S}{The number of Monte Carlo samples to draw from the \code{R-INLA}
approximate posterior distribution over the latent field.}
}
\value{
A scalar log density score.
}
\description{
Compute log density score from `R-INLA} model at a single held-out
point.
}
\examples{
fit <- m1_inla(mw)
eval_inla_model(mw, fit, i = 1, S = 1000)
}
