% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison_stan.R
\name{eval_stan_model}
\alias{eval_stan_model}
\title{Compute log density score from \code{stan} model at a single held-out point.}
\usage{
eval_stan_model(sf, fit, i)
}
\arguments{
\item{sf}{A simple features data frame.}

\item{fit}{A fitted \code{stan} model.}

\item{i}{The index of the held-out data point to predict on. Should be an
integer in the range \code{1:nrow(sf)}.
approximate posterior distribution over the latent field.}
}
\value{
A scalar log density score.
}
\description{
Compute log density score from \code{stan} model at a single held-out point.
}
