% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fck.R
\name{fck_stan}
\alias{fck_stan}
\title{Fit Centroid MVN Small Area Estimation model using \code{stan}.}
\usage{
fck_stan(
  sf,
  control = "mean",
  bym2 = FALSE,
  nsim_warm = 100,
  nsim_iter = 1000,
  kernel = matern,
  ...
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{control}{How should the length-scale be computed? \code{"mean"} sets the covariance
of the average distance between centroids to be 0.01.}

\item{bym2}{Logical indicating if the centroid spatial random effects should be convoluted
with unstructured IID noise, defaults to \code{FALSE}.}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}. Kernel hyper-parameters
are fixed.
}
\examples{
fck_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
