% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_model}
\alias{fit_model}
\title{Function to fit single model.}
\usage{
fit_model(sf, fn, ...)
}
\arguments{
\item{sf}{A simple features data frame.}

\item{fn}{A fitting function.}
}
\value{
A fitted model.
}
\description{
Function to fit single model.
}
