% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{get_scale}
\alias{get_scale}
\title{Compute scale of a precision matrix using \code{R-INLA}.}
\usage{
get_scale(Q, constraint = list(A = matrix(1, 1, nrow(Q)), e = 0))
}
\arguments{
\item{Q}{A (square, symmetric) precision matrix.}

\item{constraint}{A list with arguments \code{A} and \code{e} which imposes the constraint \code{Au = e}
(where the precision of \code{u} is \code{Q}). See the \code{?INLA::f} argument \code{extraconstr}.
If \code{constraint} is the default value then this is a sum-to-zero constraint.}
}
\description{
Takes the \code{\link{riebler_gv}} of the inverse of the precision.
If the precision is singular, then the generalised inverse is used as implemented by \code{INLA::inla.qinv},
and by default it is assumed that the graph is fully connected.
}
\examples{
nb <- neighbours(mw)
Q <- nb_to_precision(nb)
get_scale(Q)
}
