% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/held_out_metrics.R
\name{held_out_metrics}
\alias{held_out_metrics}
\title{Compute forecast assessments for \code{R-INLA} model at a single held-out point.}
\usage{
held_out_metrics(fit, sf, i, S = 4000)
}
\arguments{
\item{fit}{Fitted model.}

\item{sf}{A simple features data frame.}

\item{i}{The index (in \code{1:nrow(sf)}) of the held-out point.}

\item{S}{The number of Monte Carlo samples to draw from the approximate posterior.}
}
\description{
Compute forecast assessments for \code{R-INLA} model at a single held-out point.
}
