% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiv_surveys}
\alias{hiv_surveys}
\title{HIV Survey Data.}
\format{
A simple features data frame with 2435 rows and 13 variables:
\itemize{
\item \code{survey_id}
\item \code{gid_0}
\item \code{gid_1}
\item \code{gid_2}
\item \code{name_0}
\item \code{name_1}
\item \code{name_2}
\item \code{n_cluster}
\item \code{n_obs}
\item \code{est}
\item \code{se}
\item \code{y}
\item \code{geometry}
}
}
\source{
Jeff Eaton
}
\usage{
hiv_surveys
}
\description{
A data frame containing HIV survey data from 38 surveys across 19 African countries.
}
\keyword{datasets}
