% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{info_criteria}
\alias{info_criteria}
\title{Compute DIC, WAIC and CPO for either \code{stan} or \code{R-INLA} models.}
\usage{
info_criteria(fit)
}
\arguments{
\item{fit}{A fitted \code{stan} or \code{R-INLA} model.}
}
\description{
Wrapper function for \code{stan_info_criteria} and \code{inla_info_criteria}.
}
