% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison_inla.R
\name{inla_info_criteria}
\alias{inla_info_criteria}
\title{Compute DIC, WAIC and CPO for \code{R-INLA} models.}
\usage{
inla_info_criteria(fit)
}
\arguments{
\item{fit}{A fitted \code{R-INLA} model.}
}
\value{
A list containing the DIC, WAIC and CPO.
}
\description{
Compute DIC, WAIC and CPO for \code{R-INLA} models.
}
