% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{lobo}
\alias{lobo}
\title{Create spatial leave-one-block-out training data sets.}
\usage{
lobo(sf, remove_cols = c("y"))
}
\arguments{
\item{sf}{A simple features data frame.}

\item{remove_cols}{A vector of named columns which are to have entries
replaced by \code{NA} in the training data sets. Defaults to
\code{c("y")}.}
}
\value{
A list of \code{nrow(sf)} training set lists.
Each training set list contains:
\itemize{
\item \code{data} The training data set with left-out entries.
\item \code{centre} The index of the central held-out region.
\item \code{held_out} The indices of all held-out regions.
}
}
\description{
In leave-one-block-out (LOBO) cross-validation, in each fold one spatially
contigous block is left out in order to be predicted upon.
This function creates a list of training data sets where the \code{i}th item
leaves out region \code{i} and its neighbours.
}
\examples{
lobo(mw, remove_cols = c("y", "est"))
}
