% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic}
\alias{logistic}
\title{The logistic transformation.}
\usage{
logistic(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The logistic tranformation of \code{x}.
}
\description{
Compute the logistic (inverse logit) transformation, given by
\eqn{x \mapsto \exp(x) / (1 + \exp(x))}.
}
