% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_model}
\alias{lookup_model}
\title{Model lookup table.}
\format{
An object of class \code{data.frame} with 7 rows and 2 columns.
}
\usage{
lookup_model
}
\description{
Used for reference and in the function \code{\link{get_model_method}}.
}
\keyword{datasets}
