% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m0_constant.R
\name{m0_inla}
\alias{m0_inla}
\title{Fit constant Small Area Estimation model using \code{R-INLA}.}
\usage{
m0_inla(sf, verbose = FALSE)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{verbose}{Should \code{R-INLA} run in mode \code{verbose = TRUE}.}
}
\description{
Simply fits a constant (the mean). This is useful as a benchmark
for other models.
}
\examples{
fit <- m0_inla(mw)
mean <- plogis(fit$summary.fixed$mean)
}
