% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1_iid.R
\name{m1_stan}
\alias{m1_stan}
\title{Fit IID Small Area Estimation model using \code{stan}.}
\usage{
m1_stan(sf, nsim_warm = 100, nsim_iter = 1000)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}
}
\description{
Random effects are independent and identically distributed.
}
\examples{
m1_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
