% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_besag.R
\name{m2_inla}
\alias{m2_inla}
\title{Fit Besag Small Area Estimation model using \code{R-INLA}.}
\usage{
m2_inla(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\description{
Random effects have an improper conditional autoregressive (ICAR)
distribution. This is implemented by \code{R-INLA} using the
option \code{model = "besag"}.
}
\examples{
m2_inla(mw)
}
