% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_besag.R
\name{m2_stan}
\alias{m2_stan}
\title{Fit Besag Small Area Estimation model using \code{stan}.}
\usage{
m2_stan(sf, nsim_warm = 100, nsim_iter = 1000, method = "default")
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}

\item{`method`}{One of \code{"default"} or \code{"morris"}.}
}
\description{
Random effects have an improper conditional autoregressive (ICAR)
distribution with (generalised) precision matrix produced using
the \code{\link{nb_to_precision}} function with input \code{nb},
the neighbourhood structure of \code{sf}.
}
\examples{
m2_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
