% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_bym2.R
\name{m3_inla}
\alias{m3_inla}
\title{Fit BYM2 Small Area Estimation model using \code{R-INLA}.}
\usage{
m3_inla(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\description{
Fit BYM2 Small Area Estimation model using \code{R-INLA}.
}
\examples{
m3_inla(mw)
}
