% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m5_mvn-centroid.R
\name{m5_inla}
\alias{m5_inla}
\title{Fit MVN Small Area Estimation model using \code{R-INLA}.}
\usage{
m5_inla(sf, kernel = matern, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}.
}
\examples{
m5_inla(mw)
}
