% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m5_mvn-centroid.R
\name{m5_stan}
\alias{m5_stan}
\title{Fit Centroid MVN Small Area Estimation model using \code{stan}.}
\usage{
m5_stan(sf, nsim_warm = 100, nsim_iter = 1000, kernel = matern, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}. Kernel hyper-parameters
are fixed.
}
\examples{
m5_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
