% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6_mvn-sample.R
\name{m6_inla}
\alias{m6_inla}
\title{Fit Sampling MVN Small Area Estimation model using \code{R-INLA}.}
\usage{
m6_inla(sf, kernel = matern, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{sampling_covariance}}.
}
\examples{
m6_inla(mw)
}
