% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m7_mvn-centroid-bayes.R
\name{m7_stan}
\alias{m7_stan}
\title{Fit Bayesian Centroid MVN Small Area Estimation model using \code{stan}.}
\usage{
m7_stan(sf, bym2 = FALSE, nsim_warm = 100, nsim_iter = 1000)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{bym2}{Logical indicating if the spatial random effects should be convoluted
with unstructured IID noise, defaults to \code{FALSE}.}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}. Unlike \code{m5_stan}, this
version has a fully Bayesian treatment of kernel hyperparameters.
}
\examples{
m7_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
