% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m8_mvn-integrated-bayes.R
\name{m8_stan}
\alias{m8_stan}
\title{Fit Bayesian Integrated MVN Small Area Estimation model using \code{stan}.}
\usage{
m8_stan(
  sf,
  bym2 = FALSE,
  L = 10,
  type = "hexagonal",
  nsim_warm = 100,
  nsim_iter = 1000
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{bym2}{Logical indicating if the spatial random effects should be convoluted
with unstructured IID noise, defaults to \code{FALSE}.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{type}{String passed to the \code{type} argument of \code{sf::st_sample}, defaults to \code{"hexagonal"}}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}
}
\description{
Fit Bayesian Integrated MVN Small Area Estimation model using \code{stan}.
}
\examples{
m8_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
