% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_intervals.R
\name{marginal_intervals}
\alias{marginal_intervals}
\alias{marginal_intervals.inla}
\alias{marginal_intervals.stanfit}
\title{Intervals for posterior marginals.}
\usage{
marginal_intervals(fit, ...)

\method{marginal_intervals}{inla}(fit)

\method{marginal_intervals}{stanfit}(fit, parameter)
}
\arguments{
\item{fit}{Fitted model.}
}
\description{
Intervals for posterior marginals.
}
