% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{matrix_plot}
\alias{matrix_plot}
\title{Produce a \code{ggplot2} visualisation of a matrix.}
\usage{
matrix_plot(M)
}
\arguments{
\item{M}{A matrix.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Produce a \code{ggplot2} visualisation of a matrix.
}
\examples{
M <- matrix(1:9, nrow = 3, ncol = 3)
matrix_plot(M)

matrix_plot(border_precision(mw))
}
