% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mw}
\alias{mw}
\title{Malawi DHS 2015 Survey Data.}
\format{
A simple features data frame with 28 rows and 13 variables.
}
\source{
Jeff Eaton
}
\usage{
mw
}
\description{
\code{hiv_surveys} subsetted to just the DHS 2015 survey in Malawi.
}
\keyword{datasets}
