% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{nb_to_precision}
\alias{nb_to_precision}
\title{Create Besag precision matrix from \code{nb}.}
\usage{
nb_to_precision(nb)
}
\arguments{
\item{nb}{A neighbourhood list object.}
}
\description{
Creates the Besag model ICAR precision matrix \eqn{Q} with entries
\eqn{Q_{ij} = -1} if \eqn{i \sim j} (where \eqn{\sim} is an adjacency operator),
\eqn{Q_{ii}} is the number of neighbours of area \eqn{i} and \eqn{Q_{ij} = 0} otherwise.
}
\examples{
nb <- neighbours(mw)
nb_to_precision(nb)
}
