% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{report_interval}
\alias{report_interval}
\title{Mean, standard deviation and posterior interval for \code{stan} and
\code{R-INLA} models.}
\usage{
report_interval(fit, type, name = NULL)
}
\arguments{
\item{fit}{Either a \code{stan} or \code{R-INLA} model.}

\item{type}{One of \code{c("stan", "inla")}.}

\item{name}{The name of the parameter of interest
(only for \code{type = "stan"}).}
}
\value{
A data frame with containing the posterior mean, standard deviation
together with lower (2.5\%) and upper (97.5\%) quantiles.
}
\description{
Mean, standard deviation and posterior interval for \code{stan} and
\code{R-INLA} models.
}
\examples{
fit1 <- m1_inla(mw)
fit2 <- m1_stan(mw, nsim_warm = 0, nsim_iter = 100)
report_interval(fit1, type = "inla")
report_interval(fit2, type = "stan", name = "rho")
}
