% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_structures.R
\name{scale_gmrf_precision}
\alias{scale_gmrf_precision}
\title{Scales the precision matrix of a Gaussian Markov random field.}
\source{
From \href{https://github.com/mrc-ide/naomi/blob/master/R/car.R}{code} by Jeff Eaton.
}
\usage{
scale_gmrf_precision(Q, A = matrix(1, 1, nrow(Q)))
}
\arguments{
\item{Q}{A (square, symmetric) precision matrix.}
}
\description{
Implements the same thing as \code{INLA::inla.scale.model}.
}
\examples{
nb <- neighbours(mw)
Q <- nb_to_precision(nb)
scale_gmrf_precision(Q)
}
