% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison_stan.R
\name{stan_info_criteria}
\alias{stan_info_criteria}
\title{Compute DIC and WAIC for \code{stan} models.}
\usage{
stan_info_criteria(fit)
}
\arguments{
\item{fit}{A fitted \code{stan} model.}
}
\value{
A list containing the DIC and WAIC.
}
\description{
Compute DIC and WAIC for \code{stan} models.
}
