% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{tune_lengthscale}
\alias{tune_lengthscale}
\title{Perform lengthscale hyperparameter tuning for kernel SAE methods.}
\usage{
tune_lengthscale(sf, fun = m5_inla, grid = seq(0.05, 0.5, by = 0.025), ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{grid}{A sequence of length-scale parameters.}

\item{...}{Additional arguments to \code{fn}.}
}
\description{
Perform lengthscale hyperparameter tuning for kernel SAE methods.
}
