% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wicar.R
\name{wicar_inla}
\alias{wicar_inla}
\title{Fit weighted ICAR Small Area Estimation model using \code{R-INLA}.}
\usage{
wicar_inla(sf, verbose = FALSE)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{verbose}{Should \code{R-INLA} run in mode \code{verbose = TRUE}.}
}
\description{
Fit weighted ICAR Small Area Estimation model using \code{R-INLA}.
}
\examples{
wicar_inla(mw)
}
