% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wicar.R
\name{wicar_stan}
\alias{wicar_stan}
\title{Fit weighted ICAR Small Area Estimation model using \code{stan}.}
\usage{
wicar_stan(sf, nsim_warm = 100, nsim_iter = 1000)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}
}
\description{
Fit weighted ICAR Small Area Estimation model using \code{stan}.
}
\examples{
wicar_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
