% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_methods.R
\name{centroid_covariance}
\alias{centroid_covariance}
\title{Compute centroid kernel gram matrix.}
\usage{
centroid_covariance(sf, control = "mean", kernel = matern, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{control}{How should the length-scale be computed? \code{"mean"} sets the covariance
of the average distance between centroids to be 0.01.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Compute centroid kernel gram matrix.
}
\examples{
centroid_covariance(mw)
}
