% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_methods.R
\name{centroid_covariance}
\alias{centroid_covariance}
\title{Compute centroid kernel Gram matrix.}
\usage{
centroid_covariance(sf, kernel = matern, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Compute centroid kernel Gram matrix.
}
\examples{
centroid_covariance(mw)
}
