% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\title{Cross-validation using \code{create_folds} and evaluation using \code{held_out_metrics}.}
\usage{
cross_validate(sf, type = "LOO", fn, S = 4000, output = FALSE, ...)
}
\arguments{
\item{sf}{A simple features data frame.}

\item{type}{One of \code{"LOO"} or \code{"SLOO"}.}

\item{fn}{A model fitting function.}

\item{S}{The number of Monte Carlo samples to draw from the approximate posterior.}

\item{output}{Should \code{cv} output the training data and fitted models?}

\item{...}{Additional arguments to \code{fn}.}
}
\description{
Cross-validation using \code{create_folds} and evaluation using \code{held_out_metrics}.
}
