% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_structures.R
\name{get_scale}
\alias{get_scale}
\title{Compute scale constant using \code{R-INLA}.}
\usage{
get_scale(nb, method = "direct")
}
\arguments{
\item{nb}{A neighbourhood list object.}

\item{method}{One of \code{c("direct", "inversion")}.
I think that they should be the same.}
}
\description{
Compute scale constant using \code{R-INLA}.
}
\section{Warning}{

Have not taken into account connectedness, this is a to do.
}

\examples{
nb <- neighbours(mw)
get_scale(nb, method = "direct")
}
