% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1_iid.R
\name{m1_inla}
\alias{m1_inla}
\title{Fit IID Small Area Estimation model using \code{R-INLA}.}
\usage{
m1_inla(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\description{
Random effects are independent and identically distributed.
}
\examples{
m1_inla(mw)
}
