% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_wicar.R
\name{m4_inla}
\alias{m4_inla}
\title{Fit weighted ICAR Small Area Estimation model using \code{R-INLA}.}
\usage{
m4_inla(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\description{
Fit weighted ICAR Small Area Estimation model using \code{R-INLA}.
}
\examples{
m4_inla(mw)
}
