% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m5_mvn-centroid.R
\name{m5_inla}
\alias{m5_inla}
\title{Fit Centroid MVN Small Area Estimation model using \code{R-INLA}.}
\usage{
m5_inla(sf, control = "mean", verbose = FALSE, kernel = matern, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{control}{If not provided to the kernel, how should the length-scale be computed?
Default \code{"mean"} sets the covariance of the average distance between centroids to be 0.01.}

\item{verbose}{Should \code{R-INLA} run in mode \code{verbose = TRUE}.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{centroid_covariance}}.
}
\examples{
m5_inla(mw)
}
