% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6_mvn-integrated.R
\name{m6_inla}
\alias{m6_inla}
\title{Fit Sampling MVN Small Area Estimation model using \code{R-INLA}.}
\usage{
m6_inla(sf, control = "mean", verbose = FALSE, L = 50, kernel = matern, ...)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{control}{If not provided to the kernel, how should the length-scale be computed?
Default \code{"mean"} sets the covariance of the average distance between centroids to be 0.01.}

\item{verbose}{Should \code{R-INLA} run in mode \code{verbose = TRUE}.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{sampling_covariance}}.
}
\examples{
m6_inla(mw)
}
