% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6_mvn-sample.R
\name{m6_stan}
\alias{m6_stan}
\title{Fit Sampling MVN Small Area Estimation model using \code{stan}.}
\usage{
m6_stan(sf, nsim_warm = 100, nsim_iter = 1000, kernel = matern, ..., L = 100)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}

\item{L}{The number of Monte Carlo samples to draw from each region.}
}
\description{
Random effects have a multivariate Gaussian distribution with covariance
matrix calculated using \code{\link{sampling_covariance}}.
}
\examples{
m6_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
