% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m8_mvn-sample-bayes.R
\name{m8_stan}
\alias{m8_stan}
\title{Fit Bayesian Sampling MVN Small Area Estimation model using \code{stan}.}
\usage{
m8_stan(sf, L = 50, nsim_warm = 100, nsim_iter = 1000)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{nsim_warm}{Number of warmup samples, passed to \code{stan}.}

\item{nsim_iter}{Number of samples, passed to \code{stan}.}
}
\description{
Fit Bayesian Sampling MVN Small Area Estimation model using \code{stan}.
}
\examples{
m8_stan(mw, nsim_warm = 0, nsim_iter = 100)
}
