% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matern}
\alias{matern}
\title{Matern 3/2 covariance.}
\usage{
matern(r, l = 1)
}
\arguments{
\item{r}{A distance between two points.}

\item{l}{A lengthscale, defaults to \code{l = 1}.}
}
\value{
The Matern covariance.
}
\description{
The Matern covariance function with smoothness parameter 3/2.
}
