% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_structures.R
\name{nb_to_precision}
\alias{nb_to_precision}
\title{Create ICAR precision matrix from \code{nb}}
\usage{
nb_to_precision(nb)
}
\arguments{
\item{nb}{A neighbourhood list object.}
}
\value{
An ICAR precision matrix based upon \code{nb}.
}
\description{
This function creates the Besag model ICAR precision matrix \eqn{Q}.
\eqn{Q} has the entries \eqn{Q_{ij} = -1} if \eqn{i \sim j} (where
\eqn{\sim} is an adjacency operator), \eqn{Q_{ii}} is the number of
neighbours of region \eqn{i} and \eqn{Q_{ij} = 0} otherwise.
}
\examples{
nb <- neighbours(mw)
nb_to_precision(nb)
}
