% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_structures.R
\name{neighbours}
\alias{neighbours}
\title{Create neighbourlood list from from \code{sf} object.}
\usage{
neighbours(sf)
}
\arguments{
\item{sf}{A simple features object with some geometry.}
}
\value{
A neighbourhood list object.
}
\description{
Wrapper function for \code{spdep::poly2nb}. For singletons returns \code{NULL} in the
relevant list entry rather than \code{0}.
}
\examples{
nb <- neighbours(mw)
}
