% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_structures.R
\name{sampling_covariance}
\alias{sampling_covariance}
\title{Compute regional covariance matrix using sample-based kernel.}
\usage{
sampling_covariance(sf, kernel = matern, ..., L = 100)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{...}{Additional arguments to \code{kernel}.}

\item{L}{The number of Monte Carlo samples to draw from each region.}
}
\value{
A \code{nrow(sf)} by \code{nrow(sf)} matrix.
}
\description{
\code{sampling_covariance} draws \code{S} samples uniformly from inside each
region of \code{sf}. The kernel function \code{kernel} (with additional
arguments \code{...}) is averaged over each pair of draws between regions
to produce the entries of the covariance matrix.
}
\examples{
sampling_covariance(mw)
}
