% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_structures.R
\name{sampling_covariance}
\alias{sampling_covariance}
\title{Compute areal covariance matrix using sample-based kernel.}
\usage{
sampling_covariance(
  sf,
  control = "mean",
  L = 10,
  kernel = matern,
  type = "hexagonal",
  ...
)
}
\arguments{
\item{sf}{A simple features object with some geometry.}

\item{control}{If not provided to the kernel, how should the length-scale be computed?
Default \code{"mean"} sets the covariance of the average distance between centroids to be 0.01.}

\item{L}{The number of Monte Carlo samples to draw from each area.}

\item{kernel}{A kernel function, defaults to \code{matern}.}

\item{type}{String passed to the \code{type} argument of \code{sf::st_sample}, defaults to \code{"hexagonal"}}

\item{...}{Additional arguments to \code{kernel}.}
}
\value{
A \code{nrow(sf)} by \code{nrow(sf)} matrix.
}
\description{
\code{sampling_covariance} draws \code{S} samples uniformly from inside each
area of \code{sf}. The kernel function \code{kernel} (with additional
arguments \code{...}) is averaged over each pair of draws between areas
to produce the entries of the covariance matrix.
}
\examples{
sampling_covariance(mw)
}
