% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_structures.R
\name{scale_gmrf_precision}
\alias{scale_gmrf_precision}
\title{Scales the precision matrix of a Gaussian Markov random field.}
\source{
From \href{https://github.com/mrc-ide/naomi/blob/master/R/car.R}{code} by Jeff Eaton.
}
\usage{
scale_gmrf_precision(Q, A = matrix(1, 1, nrow(Q)))
}
\arguments{
\item{Q}{A (square, symmetric) precision matrix.}

\item{A}{See the \code{constraint} argument of \code{\link{get_scale}}.}
}
\value{
A matrix of the same dimension as \code{Q} where the entries have
been scaled according to the paper "A note on intrinsic conditional
autoregressive models for disconnected graphs" by Freni-Sterrantino,
Ventrucci and Rue.
}
\description{
Implements the same thing as \code{INLA::inla.scale.model}.
}
\examples{
nb <- neighbours(mw)
Q <- nb_to_precision(nb)
scale_gmrf_precision(Q)
}
